requirements.txt
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg2==1.19.4
fcm-django==1.0.8
firebase-admin==5.2.0
jsonfield==3.1.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
requests==2.27.1
sqlparse==0.4.2